+++ scripts/resetctdb.sh
@@ -41,6 +41,7 @@ collect_vars() {
   # handle flags
   FORCE=false
   VERBOSE=false
+
   while [[ $# -gt 0 ]]; do
     case "$1" in
       --force) FORCE=true ;;
@@ -74,16 +75,21 @@ main() {
 
   if [ -z ${REPLY+x} ] || [[ $REPLY =~ ^[Yy]$ ]]
   then
-      echo "Resetting DB..."
-      mysql "${FLAGS[@]}" -e "DROP DATABASE IF EXISTS ${MYSQL_DATABASE};" || \
-        die "Error: Failed to drop database '${MYSQL_DATABASE}'."
-      mysql "${FLAGS[@]}" -e "CREATE DATABASE ${MYSQL_DATABASE};" || \
-        die "Error: Failed to create database '${MYSQL_DATABASE}'."
-      mysql "${FLAGS[@]}" -e "CREATE USER IF NOT EXISTS ${MYSQL_USER}@'${MYSQL_USER_HOST}' IDENTIFIED BY '${MYSQL_PASSWORD}';" || \
-        die "Error: Failed to create user '${MYSQL_USER}@${MYSQL_USER_HOST}'."
-      mysql "${FLAGS[@]}" -e "GRANT ALL ON ${MYSQL_DATABASE}.* TO ${MYSQL_USER}@'${MYSQL_USER_HOST}'" || \
-        die "Error: Failed to grant '${MYSQL_USER}' user all privileges on '${MYSQL_DATABASE}'."
-      echo "Reset Complete"
+    echo "Resetting DB..."
+    mysql "${FLAGS[@]}" -e "DROP DATABASE IF EXISTS ${MYSQL_DATABASE};" || \
+      die "Error: Failed to drop database '${MYSQL_DATABASE}'."
+    mysql "${FLAGS[@]}" -e "CREATE DATABASE ${MYSQL_DATABASE};" || \
+      die "Error: Failed to create database '${MYSQL_DATABASE}'."
+    mysql "${FLAGS[@]}" -e "CREATE USER IF NOT EXISTS ${MYSQL_USER}@'${MYSQL_USER_HOST}' IDENTIFIED BY '${MYSQL_PASSWORD}';" || \
+      die "Error: Failed to create user '${MYSQL_USER}@${MYSQL_USER_HOST}'."
+    mysql "${FLAGS[@]}" -e "GRANT ALL ON ${MYSQL_DATABASE}.* TO ${MYSQL_USER}@'${MYSQL_USER_HOST}';" || \
+      die "Error: Failed to grant '${MYSQL_USER}' user all privileges on '${MYSQL_DATABASE}'."
+    mysql "${FLAGS[@]}" -e "FLUSH PRIVILEGES;" || \
+      die "Error: Failed to flush privileges."
+    mysql "${FLAGS[@]}" -D ${MYSQL_DATABASE} < ${CT_GO_PATH}/trillian/ctfe/storage/mysql/schema.sql || \
+      die "Error: Failed to import schema in '${MYSQL_DATABASE}' database."
+
+    echo "Reset Complete"
   fi
 }