+++ trillian/ctfe/storage/storage.go
@@ -17,6 +17,11 @@ package storage
 
 import (
 	"context"
+	"errors"
+	"strings"
+
+	"github.com/google/certificate-transparency-go/trillian/ctfe/configpb"
+	"github.com/google/certificate-transparency-go/trillian/ctfe/storage/mysql"
 )
 
 // IssuanceChainStorage is an interface which allows CTFE binaries to use different storage implementations for issuance chains.
@@ -26,4 +31,20 @@ type IssuanceChainStorage interface {
 
 	// Add inserts the key-value pair of issuance chain.
 	Add(ctx context.Context, key []byte, chain []byte) error
+}
+
+// NewIssuanceChainStorage returns nil for Trillian gRPC or mysql.IssuanceChainStorage when MySQL is the prefix in database connection string.
+func NewIssuanceChainStorage(ctx context.Context, backend configpb.LogConfig_IssuanceChainStorageBackend, dbConn string) (IssuanceChainStorage, error) {
+	switch backend {
+	case configpb.LogConfig_ISSUANCE_CHAIN_STORAGE_BACKEND_TRILLIAN_GRPC:
+		return nil, nil
+	case configpb.LogConfig_ISSUANCE_CHAIN_STORAGE_BACKEND_CTFE:
+		if strings.HasPrefix(dbConn, "mysql") {
+			return mysql.NewIssuanceChainStorage(ctx, dbConn), nil
+		} else {
+			return nil, errors.New("failed to initialise IssuanceChainService due to unsupported driver in CTFE storage connection string")
+		}
+	}
+
+	return nil, errors.New("unsupported issuance chain storage backend")
 }